refactor(self_relaunch_semaphore/semaphore): merge inline tests into tests.rs (#1266)
Inline #[cfg(test)] mod tests (lines 253-412, ~160 LOC) in semaphore.rs is
moved into the existing sibling tests.rs as a nested submodule
'semaphore_inline' to scope its helpers and avoid name collisions with the
existing top-level tests in tests.rs (both blocks have similar test names
like leader_state_json_round_trip, etc).

- src/self_relaunch_semaphore/semaphore.rs: 412 → 252 LOC (under 400-line rule)
- src/self_relaunch_semaphore/tests.rs: 282 → 443 LOC

Pure refactor: zero logic changes, all 39 self_relaunch_semaphore tests pass.

Closes 1/70 of #1266 (after #1285-#1289).
